app/requirements.txt
Flask==2.2.2
pytest==7.4.0
Werkzeug==2.3.7
azure-cosmos==4.3.0
